[tests/tests.c] Fixed a potential integer overflow from r12053 with GMP_CHECK_RANDOMIZE=1, in particular under a 32-bit Linux.
[From SVN r12054 (trunk)]

+++ tests/tests.c
@@ -398,7 +398,7 @@ tests_rand_start (void)
 #ifdef HAVE_GETTIMEOFDAY
           struct timeval  tv;
           gettimeofday (&tv, NULL);
-          seed = 1000000 * tv.tv_sec + tv.tv_usec;
+          seed = 1000000 * (unsigned long) tv.tv_sec + tv.tv_usec;
 #else
           time_t  tv;
           time (&tv);
